test: 테스트 코드 작성 (#70)
- 테스트 코드를 열심히 작성하였습니다

+++ winwin-be-api/src/main/java/com/dpm/winwin/api/post/dto/response/MyPagePostResponse.java
@@ -3,11 +3,16 @@ package com.dpm.winwin.api.post.dto.response;
 import com.dpm.winwin.domain.dto.post.MyPagePostDto;
 import java.util.List;
 
-public record MyPagePostResponse(Long id, String title, String subCategory, boolean isShare, List<String> takenTalents,
-                                 Long likes) {
+public record MyPagePostResponse(Long id,
+                                 String title,
+                                 String subCategory,
+                                 boolean isShare,
+                                 List<String> takenTalents,
+                                 int likes) {
 
     public static MyPagePostResponse of(MyPagePostDto myPagePostDto) {
         return new MyPagePostResponse(myPagePostDto.id(), myPagePostDto.title(),
-            myPagePostDto.subCategory(), myPagePostDto.isShare(), myPagePostDto.takenTalents(), myPagePostDto.likes());
+            myPagePostDto.subCategory(), myPagePostDto.isShare(), myPagePostDto.takenTalents(),
+            myPagePostDto.likes());
     }
 }
